rename view holders
